backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
hyperliquid-python-sdk==0.20.0
pydantic>=2.8.0
eth-account==0.10.0
websockets>=12.0
requests>=2.31.0

